FreeBSD: One fix for 64-bit system
Convert pthread_self()'s return value to long int to prevent an error
in compilation.

+++ src/util/light_process.cpp
@@ -414,7 +414,7 @@ void LightProcess::runShadow(int fdin, int fdout) {
 }
 
 int LightProcess::GetId() {
-  return (int)pthread_self() % g_procs.size();
+  return (long int)pthread_self() % g_procs.size();
 }
 
 FILE *LightProcess::popen(const char *cmd, const char *type,
